user_service/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.23.2
SQLAlchemy>=2.0.22,<3.0
asyncpg>=0.29.0
alembic==1.15.0
python-dotenv==1.0.1
pydantic[email]==2.12.3
pydantic-settings==2.11.0
redis>=5.0.0
passlib==1.7.4
python-jose==3.5.0
argon2-cffi>=21.3.0